refactor: 컴포넌트들을 components 폴더로 이동
