+++ README.md
@@ -63,4 +63,3 @@ FE 레벨2 페이먼츠 미션
 - 유효기간 입력이 완료되면, console에 빈 값이 뜸.
 - 커스텀 훅에서 대부분의 일들을 처리하고 cardNumbers와 같은 곳들에서는 가볍게 처리하도록 수정 (중복되는 내용이 많음 예를 들어, useCardNumbers와 cardNumbers에 모두 handleCardNumberChange가 있음.)
 - 4글자 입력 시 에러 뜨고, 지웠을 때 에러 사라지도록 하기
-- 상수 메시지 한번에 다쓰기 (~~ 입력해주세요는 따로 안빼도 됨)
