+++ src/App.css
@@ -33,3 +33,14 @@ body {
   display: flex;
   flex-direction: column-reverse;
 }
+
+.checkButton-container {
+  padding: 18px 0 8px 0;
+}
+
+.check-button {
+  background-color: var(--color-black);
+  color: var(--color-white);
+  width: 100%;
+  height: 52px;
+}
