+++ src/components/CheckButton.tsx
@@ -1,0 +1,7 @@
+export default function CheckButton() {
+  return (
+    <button className="check-button">
+      <span>확인</span>
+    </button>
+  );
+}
