+++ src/hooks/useCvcNumber/useCvcNumberState.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { isValidLength } from "../../validation/validate";
 import { updateArrayAt } from "../../utils/updateArrayAt";
 
-const USE_CVC_NUMBER = {
+const CVC_RULE = {
   INVALID_LENGTH_ERROR: "CVC 번호는 3자리 숫자여야 합니다.",
   MAX_CVC_LENGTH: 3,
 } as const;
@@ -18,10 +18,10 @@ export default function useCvcNumber(): UseCvcNumberReturn {
   const [cvcNumbersError, setError] = useState([""]);
 
   const cvcNumbersValidate = (value: string) => {
-    if (!isValidLength(value.length, USE_CVC_NUMBER.MAX_CVC_LENGTH)) {
+    if (!isValidLength(value.length, CVC_RULE.MAX_CVC_LENGTH)) {
       updateArrayAt({
         array: cvcNumbersError,
-        newValue: USE_CVC_NUMBER.INVALID_LENGTH_ERROR,
+        newValue: CVC_RULE.INVALID_LENGTH_ERROR,
         index: 0,
         setState: setError,
       });
@@ -29,7 +29,7 @@ export default function useCvcNumber(): UseCvcNumberReturn {
       return;
     }
     const newCvcNumbers = [...cvcNumbers];
-    newCvcNumbers[0] = value.slice(0, USE_CVC_NUMBER.MAX_CVC_LENGTH);
+    newCvcNumbers[0] = value.slice(0, CVC_RULE.MAX_CVC_LENGTH);
     setCardCvcNumbers(newCvcNumbers);
 
     updateArrayAt({
